null operator

+++ commplete_commerce_app/lib/view/widgets/item_details.dart/item_price.dart
@@ -11,7 +11,7 @@ class ItemPrice extends StatelessWidget {
     this.itemDiscount = 0,
   });
 
-  final int? itemPrice;
+  final double? itemPrice;
   final int? itemDiscount;
 
   @override
